+++ lib/modules/settings/modules/backuprestore/widgets/backup_tile.dart
@@ -17,14 +17,7 @@ class SettingsBackupRestoreBackupTile extends StatelessWidget {
             if(_values[0]) {
                 String data = Export.export();
                 String encrypted = LunaEncryption.encrypt(_values[1], data);
-                if(encrypted != Constants.ENCRYPTION_FAILURE) {
-                    await LunaFileSystem.exportConfigToFilesystem(encrypted);
-                    showLunaSuccessSnackBar(
-                        context: context,
-                        title: 'Backed Up',
-                        message: 'Backups are located in the application directory',
-                    );
-                }
+                if(encrypted != Constants.ENCRYPTION_FAILURE) await LunaFileSystem.exportConfigToFilesystem(encrypted);
             }
         } catch (error, stack) {
             LunaLogger.error('SettingsGeneralConfiguration', '_backup', 'Backup Failed', error, stack);